requirements.txt
APScheduler==3.11.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
soupsieve==2.6
tzlocal==5.2
urllib3==2.2.3